refactor create and edit
