+++ client/src/components/elements/EntityCreate.js
@@ -2,11 +2,13 @@ import React, { Component } from 'react';
 import PropTypes from 'prop-types';
 import { withStyles } from '@material-ui/core/styles';
 import Typography from '@material-ui/core/Typography';
-
+import Button from './Button';
+import ProgressButton from './ProgressButton';
 import DocumentTitle from './DocumentTitle';
 import EntityDirectoryButton from './EntityDirectoryButton';
 import ErrorBoundary from './ErrorBoundary';
 import { Page, PageLeft, PageMain } from './Page';
+import TextField from './TextField';
 import ValidationError from './ValidationError';
 
 class EntityCreate extends Component {
@@ -15,8 +17,14 @@ class EntityCreate extends Component {
       classes = {},
       entityType,
       errorMessage = null,
+      withFieldData,
+      dirtinessContext,
       renderForm,
+      buttonResetProps,
+      buttonSubmitProps,
     } = this.props;
+
+    const ReasonField = withFieldData(TextField, 'provenance/why');
 
     return (
       <DocumentTitle title={`Create ${entityType}`}>
@@ -31,7 +39,31 @@ class EntityCreate extends Component {
               Add {entityType}
             </Typography>
             <ValidationError {...errorMessage} />
-            {renderForm ? <ErrorBoundary>{renderForm()}</ErrorBoundary> : null}
+            {renderForm ? (
+              <ErrorBoundary>
+                {renderForm()}
+                {dirtinessContext(({ dirty }) =>
+                  dirty ? (
+                    <ReasonField
+                      label="Reason"
+                      helperText={`Why do you create this gene?`}
+                    />
+                  ) : null
+                )}
+              </ErrorBoundary>
+            ) : null}
+            <div className={classes.actions}>
+              <Button variant="raised" {...buttonResetProps}>
+                Reset
+              </Button>
+              <ProgressButton
+                variant="raised"
+                color="secondary"
+                {...buttonSubmitProps}
+              >
+                Create
+              </ProgressButton>
+            </div>
           </PageMain>
         </Page>
       </DocumentTitle>
@@ -42,8 +74,12 @@ class EntityCreate extends Component {
 EntityCreate.propTypes = {
   classes: PropTypes.object.isRequired,
   entityType: PropTypes.string.isRequired,
+  withFieldData: PropTypes.func.isRequired,
+  dirtinessContext: PropTypes.func.isRequired,
   errorMessage: PropTypes.string,
   renderForm: PropTypes.func,
+  buttonResetProps: PropTypes.func,
+  buttonSubmitProps: PropTypes.func,
 };
 
 const styles = (theme) => ({
@@ -57,6 +93,16 @@ const styles = (theme) => ({
     [theme.breakpoints.down('sm')]: {
       width: '100%',
       alignItems: 'stretch',
+    },
+  },
+  actions: {
+    marginTop: theme.spacing.unit * 2,
+    '& > *': {
+      marginRight: theme.spacing.unit,
+      width: 150,
+      [theme.breakpoints.down('xs')]: {
+        width: `calc(50% - ${theme.spacing.unit}px)`,
+      },
     },
   },
 });
