+++ client/src/components/elements/index.js
@@ -48,8 +48,8 @@ export { default as DocumentTitle } from './DocumentTitle';
 export { default as EntityNotFound } from './EntityNotFound';
 export { default as EntityDirectoryButton } from './EntityDirectoryButton';
 export { default as EntityCreate } from './EntityCreate';
-export { default as EntityCreateForm } from './EntityCreateForm';
 export { default as EntityEditForm } from './EntityEditForm';
+export { default as EntityEditNew } from './EntityEditNew';
 export { default as EntityDirectory } from './EntityDirectory';
 export { default as EntityProfile } from './EntityProfile';
 export { default as ErrorBoundary } from './ErrorBoundary';
